requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyodbc>=5.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
colorlog>=6.7.0
tqdm>=4.66.0
